chore: refactor CutFishBody and Gills into a single script
