--- Assets/FishFactory/Scripts/CutFishGill.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 721d196e533c8b04dabea81e3dd9bd46
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
